backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.0.1
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==41.0.4
Pillow==10.0.1
